requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.3
rq==1.15.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==1.10.13
minio==7.2.5
PyPDF2==3.0.1
python-docx==1.1.2
jinja2==3.1.3
pytest==8.1.1
httpx==0.27.0
